[FIX] sale_subscription: variant price
- Activate 'Discount on SO line' on user
- Activate Variants and the pricelist with advanced feature
- Create a product P with price 0 and available in Subscription
- Create 2 variants for P, V1 and V2
- On a pricelist PL, set the `discount_policy` to `without_discount`
- Set a fixed price for V1 and V2 on PL
- Create a new subscription with PL
- Add V1

The price remains zero.

It happens because `lst_price` is zero in this case.

The root cause is the commit 991e5088c6c26b6a: it introduces the support
for the `without_discount` behavior on the Subscriptions, but this
support is only partial. Indeed, a complete support of the feature would
require to duplicate the logic from the SO:

https://github.com/odoo/odoo/blob/d310f1c59c7fd28903d91b0f2fc46044d0cea6a0/addons/sale/models/sale.py#L1636-L1668
https://github.com/odoo/odoo/blob/d310f1c59c7fd28903d91b0f2fc46044d0cea6a0/addons/sale/models/sale.py#L1585-L1628
https://github.com/odoo/odoo/blob/d310f1c59c7fd28903d91b0f2fc46044d0cea6a0/addons/sale/models/sale.py#L1453-L1484

(this might not be exhaustive)

We should simply not support this on Subscriptions.

opw-2367845

closes odoo/enterprise#14802

X-original-commit: 307cf8a6ccb75f9b0f9d746cf1664a88f735ee80
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>