[IMP] hr_payroll: Archived employees on payslip batch
When an employee is no longer working on the company, the employee get
inactivate, archiving the employee record.

Sometimes the company must pay one or more payslips to that employee,
for example for the time worked on the current payslip period,
or for other concepts.

This case of use is very common, for example the profit sharing. Maybe
those employees or no longer working in the company.

When it is necessary to add an archived employee in a batch, it is posible
to add it, however when confirmation button is clicked, the archived
employee is deleted and the payslips is not created.

Now it is possible to add archived employees to a batch and create their
payslips.

OPW#2380649

closes odoo/enterprise#14915

X-original-commit: feca7e8f318a9c4c77184f8052d3e64303988ea2
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>