[FIX] account_reports: enable support of multi-companies on the tax report through the company selector
In some cases, users may want to do that to consolidate the reports of multiple sub-companies. To avoid additional complexity and stick to our stability policy, we add a config parameter to enable this behavior. No currency conversion is supported here, as we expected the aggregated companies to share the same main currency.

OPW 2372358

closes odoo/enterprise#14791

X-original-commit: d787df94000b593834b0db8c2f9fe3ced4834df5
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>