[FIX] web_mobile: JS and native datepickers shouldn't be both opened
When the user clicks on a {date|datetime} field in the mobile app, the
native datepicker should be presented instead of the JS Tempusdominus
datepicker.

Before this commit both are shown (the JS one partially hidden behind
the native one).

This commit fixes this by properly overriding the method
`_onInputClicked` when the native datepicker is available to prevent
from calling the `$.datetimepicker()` plugin.

Note: to introduce appropriate tests for the FieldDate and FieldDateTime
fields in the mobile application, the check for the native datepicker
availability have to be performed at runtime instead of load time ; this
allows to mock it in the tests.

closes odoo/enterprise#14836

X-original-commit: e617e19da7a4e0af38198fb3eeb2c1bab45b6d96
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>