[FIX] web_map: test MapBox token on write
Without this change, if the user somehow got the MapBox api token wrong,
the settings would not work anymore because there is an onchange each
time the settings are opened => this break the onchange which prevent to
have the correct data when res.config.settings and prevent any save.

Now we only check the token when it is being changed from current value,
and return a warning instead of an error.

note: this commit also improve the structure a bit

opw-2374033
closes #14770

closes odoo/enterprise#14892

X-original-commit: a201accce19d129cd090dd9e310955655bb24fb8
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>