[FIX] industry_fsm, industry_fsm_report, industry_fsm_sale: traceback on updates
Steps to reproduce the bug:

1.Create an automated action with the following:
       - Model: Task (project.task)
       - Active: True
       - Trigger: On Update
       - Action To Do: Create Next Activity

2. Open or create a task in Project module
3. Edit
4. Change or set customer (partner_id)

Bug:

An error message was raised.

PS:

The source of the problem is calls to write() in compute methods, which ends up in
code that does not expect new records in 'self'. Developers must use usual field
assignments or call update() instead.

opw:2376492

closes odoo/enterprise#14772

X-original-commit: 9c9866fad7b8db53a4547ae9dbce2bd9b147e115
Related: odoo/odoo#61812
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>