[FIX] web_enterprise: fix bad prop exception on home menu
When an application icon is missing from the filestore, the `apps` prop
is not valid because `web_icon_data` is empty, but the `web_icon` is
not a list of three elements as it still references an icon
(`module,path`).
The `backgroundColor` is then `undefined`, and in debug mode the prop
validation fails and the page remains blank with a stacktrace.
This could happen when the filestore is missing, for example if the
database has been restored from a dump without a filestore.

This commit makes the script check that the third part is not
`undefined` before using the `webIcon`, or fallbacks to the default icon
if it is not the case.

closes odoo/enterprise#14949

X-original-commit: f21b574addbdda9d976828b15330f384ef1e9ca1
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>