[FIX] website_event_[track_]social: adjust send push reminders behavior
If you have push notifications enabled, a button on the event allows to
manually send the configured push reminders for the event tracks.

Before this commit, the action would simply trigger the same method as the
social CRON job, which would send ALL ready social.posts, whether they are
related to the event or not.

Now, pressing the button will only send push notifications that are the ready
reminders for the tracks of this specific event.

We also took the opportunity to move the button and python action from
website_event_social to website_event_track_social where it belongs.

As this is a fix for stable, the method in website_event_social was left empty
to avoid breaking potential overrides. It will be removed in master.

Task 2360932

closes odoo/enterprise#15003

X-original-commit: e496155b9a28e550b236d7165949997e23194a7e
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>