[FIX] web_mobile: add shortcut to home screen is broken
When the user attempts to add the current app to his device's home screen,
an error is displayed.

Steps to reproduce:

- Open an app
- Go to the burger/user menu and open it
- Click on "Add to home screen" and a traceback occurs

Following the refactoring of the HomeMenu on enterprise, the menu's data
aren't exposed directly any more from the WebClient.

This commit fixes it by properly accessing them from the WebClient's
Menu.

Task ID: 2388619

X-original-commit: 5330f4c15eef30577f0d6edc99c0d6fe29168f3a