[FIX] helpdesk_{sale_}timesheet: allow_billable=True when use_helpdesk_sale_timesheet is True
Before this commit, when the user checks the 'time reinvoicing'
and saves in helpdesk.team, if the helpdesk_sale_timesheet is not
installed, the generated project will not contain the allow_billable set
to True. Because the generation of this project is made in
_init_data_create_project method when helpdesk_timesheet is loaded or in
write method of helpdesk.team defined in helpdesk_timesheet module.

This commit add _post_init_hook in helpdesk_timesheet and
helpdesk_sale_timesheet modules to generate project in helpdesk.team in
the last module installed (case when the user checks 'Time Reinvoicing'
and 'Timesheet on Ticket').
If the user checks only 'Timesheet on Ticket' and saves the project has
right generated with allow_billable=False. After that, if the user
checks 'Time Reinvoicing', then we set the allow_billable in this
generated project to True.

task-2388500

X-original-commit: f56e2f75e476601d1e49e356578824e32d158bf6