[FIX] web_studio: Boolean active in report editor
In the report editor, BooleanField is an Owl component but would:

- never be isMounted because it was not in DOM when `mount` was used

- not send expected structure of event to AbstractEditComponent because
  _onDirectiveChange was expecting event from old type of widget.

With this changeset, we add special cases for Owl component so clicking
on a checkbox (eg. "Display only the date" option for a datetime field)
do what is expected (enable or disable the option).

Without the change, the added test fails because the assertion:

 "Toggling no_marker checkbox should change the option value"

is never run.

opw-2381667
closes #14862

closes odoo/enterprise#14905

X-original-commit: 93ccfa2d9ec34f4f528ab8afe91e9d1af8d9b4da
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>