[IMP] l10n_be_hr_payroll: Update private car reimbursement up to 2020
Generally this is not mandatory
See: https://emploi.belgique.be/fr/themes/remuneration/intervention-de-lemployeur-dans-les-frais-de-deplacement-domicile-lieu-de
But this is the case for the CP200
See: https://www.sfonds200.be/fonds-social/infos-sectorielles/frais-de-transport/prive-2020

The test modification in which we change the basic salary could seem odd, but
it's valid. This sample contract has extra-legal holidays. As we take the
wage_with_holidays that directly depends on the final yearly costs, and higher
reimbursement implies a higher yearly employer costs, and then the salary is
a little bit lower.

closes odoo/enterprise#14889

X-original-commit: dddbe1a36ff121a7e288616e89c5c981a1ac7144
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>