[IMP] stock_barcode: auto-start draft inventories
Currently when opening a draft inventory in the barcode app, it stays in
draft and there's no way to start it without going into the backend form
view. This commit makes it so when a user clicks on a draft inventory in
the barcode app then it will start the inventory.

Note it is possible to reset an In Progress inventory back to draft
within the barcode app.

closes odoo/enterprise#14667

Task: 2336455
Related: odoo/upgrade#1839
Related: odoo/odoo#59834
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>