[FIX] helpdesk_sale_timesheet: update sale order creation in ticket
Before this commit, we can create a sale order, if the allow_billable in
the project linked in the ticket is set to False. Because, before the
allow_billable=False means the pricing type is equal to "Task rate". But
now it's not the case, it means that the project can't be invoiced.
Moreover, when the user wants to create a sale order from a ticket in
which the pricing type is equal to "project rate" or "employee rate" in
the project linked, then the quantity in the first sale order line stay
to 0. It's because the ticket is not linked to a task and then the
timesheet is not linked in the first sale order line of the sale order
created.

In this commit, we change the condition for the visibility of the
"create sale order" button to fit with the configuration of the project
linked to the ticket. And we add a context key called
"timesheets_on_ticket" to give to the wizard for the creation of the
sale order, the timesheets in the ticket. With this key, if the
project in the ticket doesn't contain a task then the timesheets in the
ticket are correctly linked in the first sale order line of the new sale
order.

task-2388500

X-original-commit: 955be0f37ca8a1f13757c4f79832a3f83755c7a4