[FIX] iot: Ingenico payments could randomly fail
The CRC value put into the MDC tag of a message was not being padded to
the right length. It should have been padded to 8 characters but was
only being padded to 2. This means that if you had a CRC like f744682,
no padding would happen to 0f744782, and an odd-length exception would
get thrown by unhexlify. This could for example disrupt the sending of a
 TransactionRequest, which would cause a payment request to fail.

closes odoo/enterprise#15165

X-original-commit: 77c58b8d0d7f47a1fa80328c20b4a671040665b2
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>