[FIX] payment_sepa_direct_debit: initialize signature control when modal is shown
Behavior prior to this commit:

- if "SEPA Direct Debit" is the only payment acquirer activated, it will
be pre-selected as payment method.  But when paying an invoice online,
the signature box is not initially shown when the payment popup shows
up.  Instead the user has to click on the payment option again, to get
the form to reload.
- this does not happen when paying on the ecommerce site (because the
payment form is not shown in a modal there)

Behavior after this commit:

- when the modal is shown, if the signature control appears to have not
been initialized yet, it will be re-initialized (recalculating its size
in the process)

Note:

- we use `resetSignature`, rather than `resizeSignature`, because the
latter does not properly reset the control (the gray signature line
for example) in case there is a "large" resize (in this instance,
resizing from 0)
- this code is copied from the portal_signature code (in 27fca749af30)

opw-2389869

closes odoo/enterprise#14947

X-original-commit: 9d8b81164b357c40ce46fe66655519b758600338
Signed-off-by: Nicolas Galler <ngaller@users.noreply.github.com>