[FIX] helpdesk_timesheet: remove the warning messages in ticket
When the user changes the task of a ticket in Helpdesk App. We displayed
a warning message about the impact of this changes.
This warning message was useful to inform the user that is a change
in the behaviour compared to the previous version of Odoo.
But each time, the warning message appears and it can be boring for
the user. That's why we remove the warning message.

task-2388500

X-original-commit: 2ef9bc37301d1c3e4207bef11d3ace2ca4ffc95a