[FIX] event_barcode: fix registration desk back button
When using the "back" button in the event barcode scanner, when coming from the
"Registration Desk" menu, the user will be redirected to the event app but the
whole top menu disappears.
It makes it difficult to find your way into the other apps as you are "stuck"
on this view.

This commit fixes that behavior using a workaround that defines an entire new
action from scratch instead of using an existing one, because it seems not
correctly supported by the web client.

Task 2380205

closes odoo/enterprise#15293

X-original-commit: 5e88fbbfe6fdc11299b7d0b162a9adb7850babba
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>