[IMP] marketing_automation: improve label for an option of trigger_type
While adding an activity in the workflow of any marketing automation campaign,
users have to choose a trigger_type. In this selection field, we have one
option labeled as 'beginning of the campaign', which makes it look like as if
it is X hours(or selected interval) after the campaign started, while it is
instead X hours after the participants enters their own workflow.

This commit improves the behavior by changing the label for that option
from 'beginning of the campaign' to 'beginning of the workflow' and
thus avoiding the misinterpretation.

closes odoo/enterprise#14926

Task-id: 2224416
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>