[IMP] planning: detect conflicts by taking into account the allocated_percentage
Before this commit, the slot conflicts were only calculated by detecting overlap start_datetime and end_datetime.
This commit takes into account the allocated percentage and prevent their sum to be > 100.

TLDR: false positives are detected.

The correct way to detect conflicts would be to check the number of allocated_hours of overlapping slots and compare it to the slots range (min start_datetime to max end_datetime).
We don't have enough information to do it properly (in one SQL query without calling python methods).

Preventing the sum of allocated_percentage to be > 100 is only 100% correct when the conflicting slots starts at the exact same time and end at the same time too.
When they overlap during a fraction of their duration, the percentage could be higher and the only way to know is to use the "exact method".

closes odoo/enterprise#14743

Taskid: 2310218
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>