[IMP] helpdesk_sale_timesheet: remove 'create sale order' button
In Helpdesk App, we remove the 'create sale order' button in
helpdesk.ticket, because we want to add a sale order line field in
ticket to linked the timesheets in ticket which aren't invoiced to the
sale order line. With this newest field in ticket, we should not need
the button removed in this commit.

task-2424382