[FIX] social,website_event_track_social: improve ACLs of event and social bridge
Currently, some issues exist if you try to create event tracks when not having
social groups if push notifications are enabled in the settings.

To avoid this, we now display the push reminder field all the time on the
event.track form view and allow event users to enable/disable it freely.
Subsequent social models search/create/write are sudo'ed to avoid ACL issues.

However, you still need to have necessary social groups to be able to manually
edit the push reminder associated to the track.
This means that the flow of creating/updating tracks for event users should be
eased and subsequent modifications will be reflected to the created event
tracks.

A security test was added to ensure this behavior.

Task 2382683

closes odoo/enterprise#15283

X-original-commit: bfc6628dc1ae6fda16e01c2cacf83fa64cbcf566
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>