[IMP] mail_mobile: support for mobile notification when email incoming
Before this commit, when a user replied with an email ocn ignore them.

After this commit, all user having the Odoo Mobile App ans the option
"Handle in Odoo" will get a notification when an user reply by email to
an email.

Steps to reproduce:
* configure a in/out mail server
* send a mail to a "customer"
* reply on the email received
=> no notification (BUG)

Note
We use html2plaintext instead of html2text to reflect more the odoo code

Task ID: 2180211
opw-2172858

closes odoo/enterprise#7909

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>