[FIX] l10n_be_hr_contract_salary: Create a new bike from configurator
Currentlym When the vehicle model is a type bike and you have select that
bike from salary configurator the new vehilce for bike is not created.

We already have configuration(boolean) field on contract for new car
which will allow to create new car. but do not have configuration for
bike and we can not add new field in stable version.

so whenever user select the bike model we should allow to create it
without any configuration like car.

closes odoo/enterprise#15813

Task-id: 2372530
X-original-commit: 674ed4a369bcff78a82980f9ca0dba8a71576e94
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>