[FIX] web_enterprise: avoid traceback on touch event
When performing a "touch" on some elements (e.g. a list view inside a
form view) on mobile, a traceback is displayed with the error "Cannot
read property 'length' of undefined".

In this context, the "undefined" is the property "this.selection" in the
ListRender. It happens due to an unintentional override of the event
handler methods used for the "touchstart" and "touchend" events in both
BasicRenderer and ListRenderer, making the first one calling the second
one event handler. Those were introduced respectively in odoo/enterprise@4ee338a1de4ab0b6ae584b3c6368c94a0e62e065
and odoo/enterprise@ab8932f8958be1329f28af0d2ae01c6643af8ac6 .

Steps to reproduce (on mobile):
- Open Fleet app (but not specific to this app)
- Car > Create
- Scroll to "Fiscality: Disallowed Expenses Rates" field (which is a list)
- Click on the "%" symbol => Traceback

This commit fixes it by giving more specific names to the event
handler's methods on both BasicRenderer and ListRenderer.

closes odoo/enterprise#14976

Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>