[FIX] industry_fsm_sale: Prevent misconfiguration on fsm projects
Purpose
=======

prevent the user to configure fsm project with employee_rate or project_rate

Indeed, the fsm project doesn't correctly work to create sale order when
it is configured as project_rate or employee_rate.
That's why this commit allows only to configure a fsm project with task
rate.

Moreover, when the user creates or updates a project to have a fsm
project, we need to be sure that the sale_line_id field in project is not
set to respect the sql_constraint.

task-2388500

X-original-commit: 931dab8588db89a06e08fc9dd84617cc7ded3312