[FIX] delivery_{*}: default invoicing policy
Since commit https://github.com/odoo/odoo/commit/3ad4abe171e8e7e86ed0e7b7f000e734d8a2ad92
the default invoicing policy of a product is set to 'delivery'. This is
not ok for the delivery products: a delivery product is never delivered.

This is confusing to end users, because they do not realize that a
delivery line is not included in an invoice. This is especially true if
the delivery is free.

This commit sets the default policy to 'order' for master data.

opw-2387437

closes odoo/enterprise#14995

X-original-commit: d9678ba5419f9a6ade7284f4574d02faca5e0910
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>