[FIX] documents_spreadsheet: support undefined in PIVOT.HEADER formula
The formula `=PIVOT.HEADER("1", "product_id", A2)` would crash if
A2 is empty.
The result is now an empty string. This is the same behavior as the PIVOT
formula (this can be challenged, an #ERROR would probably be meaningfull).

Note that an id that does not exist formula will have an #ERROR.
e.g consider the same formula as above, with A2=99 and the product_id with
id=99 does not exists. This behavior seems inconsistent with what is stated
above. Also the error is wrongly displayed as "[object Object]". Task 2393876
should handle that.

opw 2376306

closes odoo/enterprise#15017

X-original-commit: 7527695e01e68cdb6c942e04ad6fbc11f56f8653
Related: odoo/odoo#62430
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>