[FIX] industry_fsm_sale: display the new quotation button
If the user disables the feature in the settings of Field Services, that
is 'Extra Quotations' (correponds to group_industry_fsm_quotations
field) then the button 'new quotation' is always displayed for fsm
project because the allow_quotations in the project is set to True.

Since the group_industry_fsm_quotations field implies a group named
'industry_fsm_sale.group_fsm_quotation_from_task', we add this group in
the groups attribute of the button to fix this problem.

task-2388500

X-original-commit: 6d431013e7886ba5bd250b964535b6335cab682f