[FIX] account_asset: depreciation rate is not calculated correctly on report
Behavior prior to this commit:

In some cases, the `depreciation_number_import` field is left as NULL in
the database, and thus the report does not calculate the depreciation
rate.  For example, this happens if the asset is created automatically
by posting a vendor bill with an account that is set up to create
assets.

Behavior after this commit:

Report SQL statement will handle the NULL value and treat them as 0

Note:

The setting of default values on fields is not consistent between
records created via the UI, and records created programmatically.  This
is because for records created via the UI, the `onchange` method is
called initially to retrieve defaults, which may include implicit
defaults set via the field type (in `convert_to_record`).

opw-2391268

closes odoo/enterprise#15029

X-original-commit: 163131c51026755549b8d27764351a184d73f197
Signed-off-by: Nicolas Galler <ngaller@users.noreply.github.com>