[FIX] account_reports: fix preview (pdf) printing with account groups
Go to Accounting > Configurations > Chart of Accounts
Click on an account (i.e. 400000 - Product Sales)
Add a Custom Group for this account
Go to Configuration>Financial Reports>Profit and Loss
Check 'Enable the hierarchy option'
Go to Reporting > Profit and Loss
You can see your group for the account 400000 and fold unfold it
Fold the group and click "Print preview"

The resulted PDF have the group unfolded, this because by default all
results are collected and then organized by the view. That does not
occur when printing.

The backend fix introduce a view issue: when we unfold a line that
contains a folded line, the lines inside of
the folded line were displayed. This happens with account group
hierarchies because all the lines are given at the same time and not
computed with `_get_lines`.
To fix this, we call `fold` client side once we have loaded the new
lines.

opw-2363814

closes odoo/enterprise#15420

X-original-commit: d2201f691f3a690dc0bac97f51d97d48a0731792
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
Co-authored-by: wan <wan@odoo.com>