[FIX] web_gantt: fix undeterministic issue in test
Before:
The test 'click button scale in empty gantt with sample="1"' failed
randomly. This was caused by the random generation of field values in
SampleServer which can sometimes not generate dates for today's range.

Now:
We patch SampleServer._getRandomDate in the createView to always
return the initial date.
This makes all the tests based on sample data deterministic as every
record will be in the initial date's slot.

closes odoo/enterprise#15112

X-original-commit: d6e15ea602a1ae43a8fc6aa046c1d1d40ff5c2a4
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>