[FIX] mail_mobile: fix fallback to message values in case of resend
If a message is resend  prefetch message's values in msg_vals are not
set and so we fallback to the original message.

Python Sample:
```python
a = {}
a['test'] = 'ok'
b = [a.get('test1')] or "KO"
c = [a.get('test')] or "KO"
```
the "KO" statement is never reach for b as it's value
is [None] and so not falsy.

Ref:
https://rules.sonarsource.com/python/RSPEC-5797