[FIX] l10n_mx_edi: Allow stamp with external trade
If a complement is added (External Trade for example), the prefix and
namespace must be ser in the ``Comprobante`` to avoid the next error:
```
Code : CCE999
Message : El prefijo y namespace usados para comercio exterior son
incorrectos o no están localizados a nivel cfdi:Comprobante.
```

Now is fixed.

Note: Is not the better fix, because adds the prefix in all CFDIs, but
is the same solution that here:
https://github.com/odoo/enterprise/commit/1265284c0d6d0cfa1103bb10d2b6abeb5494e3a5

closes odoo/enterprise#15116

X-original-commit: 2716dbb67074c3473791c565842d9de4450920e5
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>