[FIX] account_bank_statement_import: Allow importing bank statements with lines at zero
Sometimes, importing a Coda file fails because the bank add comments on
a bank statement by adding a line with an amount at zero.

With this change, the statement lines will have an additional filter,
which will remove lines with an amount at zero before creating the bank
statement in odoo.

task id #2403369

closes odoo/enterprise#15193

X-original-commit: d71f224f96d7f3e043bfc9ede3a175295d657be5
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>
Signed-off-by: Nicolas Viseur <vin-odoo@users.noreply.github.com>