[IMP] documents: limit the maximum upload size on documents
PURPOSE

Currently, when we add a file too large, we wait some times before we are
notified by Odoo "413 Request Entity too large" and there is no way to limit
the maximum upload size on documents.

So in this commit, added a ir.config.parameter (System Parameter, document.max_fileupload_size),
which is set 67 MB by default based on runbot ngnix limit 70MB and the
administrator can change it from the system parameter.
it will display a notification before the upload starts to notify the user that his file is too big.

task-2313367

closes odoo/enterprise#15491

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>