[IMP] hr_payroll: Added the menu 'Time Off to Approve' in Work Entries
The purpose of this task is to add a new menu "Time Off to Approve"
in Payroll that will display the leaves that have to be approved.

In this commit a menu "Time Off to Approve" is added in the Work
Entries, that display the leaves that have to be approved and have
to be reported in payslip.

closes odoo/enterprise#15097

Taskid: 2388746
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>