[FIX] l10n_mx_edi: `MemoryError` on install when lots of move lines
Installing `l10n_mx_edi` on a database with a lot of `account.move.line`
could lead to a `MemoryError`, in addition to take a lot of time (for no reason)
when the ORM was computing the fields
`l10n_mx_edi_tariff_fraction_id`
`l10n_mx_edi_umt_aduana_id`

which are related, stored, fields from `product.template`,
which are empty/`NULL` on install (no default set for the field),
and therefore the ORM was taking a lot of time and memory
for, at the end, setting `NULL` for all records.

By setting the columns in advance,
the ORM doesn't attempt to compute these fields
on install.

This is similar to 34527639873854d2be5a9bb1e3b5c8b88e52de33

opw-2369886

closes odoo/enterprise#15119

X-original-commit: d15533b0724209357e3bc3d5a4bfa390193f0752
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>