[IMP] stock_barcode_picking_batch: batch pickings wizard
This commit adds a wizard to group pickings in batch when the user want
to open a picking from the Barcode App.
If there is other pickings with the same operation type and the same
contact than the selected picking, and if those pickings should be
already processed or are planned before the selected picking, a wizard
will ask the user if they wants to start a new batch or process the
picking as usual.

task-2394799