[FIX] web_studio: fix true/false parsing in py.js
- Install Studio
- Enable debug mode
- Go to Apps
- Activate Studio & go to Reports
- Create an External Reports
- Click on Company address
The following error is raised:
"Error: Invalid Prop 'value' in component 'CustomCheckbox'"

"no_marker" option has "true" as value, which is not correctly parsed as True Boolean value.
When options are parsed with py.js, true is treated like a "Name", unlike True & False that
are "Constants".
Therefore, the parsed value of true is {expr: "true", params: undefined, isField: true},
instead of expected boolean value "true".
Then, it raises an error during validation in owl.js (only in dev mode), because value for
"no_marker" is supposed to be a Boolean but it is an object instead.

opw-2393926

closes odoo/enterprise#15156

X-original-commit: 73f7ec59ec1f6f3619d5cc1d4896e418783d0a77
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>