[ADD] hr_contract_salary_holidays : Create allocation on contract signature
We would like to add a new allocation when a HR responsible countersigns a contract.
With this brigde module between the hr_contract_salary and hr_holidays the allocation
is created once the employee's contract is signed and only needs a validation.

A Extra Time Off leave type is added to data's.

To prevent multi-companies issues, the hr_salary_holidays_auto_allocation is not set to
True as default, due to the fact that there are no Time Off Type defined at Company creation.

Security :
If a given user is the HR responsible for a Contract Template and has no rights on hr.leave.allocation
an AccessError is raised once the responsible countersigns the contract.

Task-2391381

closes odoo/enterprise#15168

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>