[FIX] web_grid: ensure t-key unicity in template
In a project, create two tasks with the same name, and create a
timesheet entry for each of those tasks. Then go the the Timesheet
app: it never loads itself.

The issue came from the template: there is a t-foreach using the
row's label as t-key. So, if two tasks have the same name, their
rows have the same keys. Moreover, n the case of this specific
view, each row has sub Components (widget="timesheet_uom").

The following happened when Owl tried to render the grid:
 - render the first row and instantiate sub components
 - render the second row (with same t-key), destroy the previously
   instantiated components and re-instantiate them
 - wait forever for the first row (and its destroyed components) to
   finish.

Task 2413371

closes odoo/enterprise#15244

X-original-commit: ae0ca952a2987974bd0eb9cc99a7e3beabbac07c
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>