[FIX] documents: rule actions priority on mobile
In the Documents's PdfManager UI on mobile, the rule's action buttons
are grouped into a global "buttons" dropdown.
Sadly the first rule's action button is both marked with 'btn-primary'
and 'btn-secondary' classes (only on mobile).

The reason was that the condition to set the 'btn-primary' class was
based on both the 'isMobile' flag and its index in the set of rules
(only first one).

This commit fixes it by removing the index-based part of the condition
(as discussed with FLL).

Found during reviewing of task 2347584

closes odoo/enterprise#15190

X-original-commit: 580ef96eb0c01f61c69d65a40185367f2597f0a3
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>