[FIX] web_enterprise: some tooltips were not shown on mobile
Before this commit, the tooltipNodes was common for all BasicRenderer,
and so when a user switch view, the tooltipNodes was filled but then
the on_detach_callback was called for the old view and cleaned the
tooltipNodes array.

After this commit, the tooltipNodes is specific for each instance.

Note that showTimer is fixed too.

Steps to reproduce:
* Open Odoo in Responsive Mode (Mobile)
* Go from App switcher to CRM
* Select a Lead
* Do a long touch on the "Tags" label
=>  no tooltip is shown (Bug)

X-original-commit: 786f4facb8f29ff7590c674b67286b1d2b04d702