[FIX] web_enterprise: don't show tooltip on mobile when not needed
Before this commit, when a element has children, the "Touch Start" event
is triggered from the children.

After this commit, the show function is only triggered when the event
element has the class 'o_user_select_none'.
This code is for a specific case when a element with children has a
tooltip.

e.g.: The progress stage in a form view
This field is rendered in a dropdown menu on Mobile, so when the "Touch Start"
event is triggered, the show method was triggered. But if the user closes
the dropdown before the end of the timer. An error will occur because the
hide method was called for a hidden element.

Snippet from tooltip.js (Bootstrap)
```javascript
    _proto.show = function show() {
      var _this = this;

      if ($(this.element).css('display') === 'none') {
        throw new Error('Please use show on visible elements');
      }
      // ...
    }
```

closes odoo/enterprise#15261

X-original-commit: 0d5fcbb49f47e5d46135ecd0ac8bb5581900fa08
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>