[IMP] web_{enterprise,studio}: add custom field button
This commit adds a "Add Custom Field" button on optional columns
dropdown list.
The goal is to let the user now he can add custom fields to a listview
with Odoo Studio, and to promote Odoo Studio when not installed.

This commit aims to add a button which opens studio from the optional
column dropdown menu when studio installed.

If studio is not yet installed, this button opens a dialog which
promotes studio with the studio youtube video. System admin can choose
to install studio directly by clicking the Install Studio button.

By clicking on this button, studio will be installed, a value will be
set in the localStorage and the page will be reloaded. The purpose of
the value set in the localStorage is to open the page in studio mode
directly after page reload and to mimify the user experience described
here after.

When Studio is installed, the buttons only toggle studio mode as the
systray button do.

--------------------------------------------------------------------
Technical comment :
For testing purposes, we must isolate the page reload function in order
to be able mock it.
--------------------------------------------------------------------

PR community : odoo/odoo#66979
PR enterprise : #15205
task-2392303

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Aaron Bohy <aab@odoo.com>