[IMP] helpdesk(_sale_timesheet): Allow to set sla based on customer & SO line
Purpose of the task is to define which customers belong to which support
category and based on their Sales order Item or customer the sla policies
should be applied.

So in this commit done the below improvements:
 - added a 'Customers' and 'Sales Order Item' m2m fields on SLA policies
 - the SLA policy should be computed accordingly on tickets:
   - if a customer is selected, the sla policy should apply on any
of his tickets (including the ones from any individual contact linked
to the company)
   - if a sales order Item is selected, the sla policy should only apply on
tickets with this SO line set as ref. sales order item
   - if the customer or SO line fields are left empty, the sla policy should
apply to any ticket without distinction.

closes odoo/enterprise#15222

Taskid: 2382672
Related: odoo/upgrade#2078
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>